+++ correctness/Edge.hpp
@@ -6,15 +6,15 @@ class Edge
 {
 public:
     /// @brief the two endpoints of the edge
-    long long first, second;
+    int first, second;
     /// @brief the two endpoints of the opposite edge
-    long long opposite_first, opposite_second;
+    int opposite_first, opposite_second;
     /// @brief iterator to the position of the edge in the list of chords
     list<Edge*>::iterator chordItrGS;
     list<Edge*>::iterator chordItrVGS;
     bool isValid;
     /// @brief  the constructor of the class
-    Edge(long long a = 0, long long b = 0, long long c = 0, long long d = 0) : first(a), second(b), opposite_first(c), opposite_second(d) {
+    Edge(int a = 0, int b = 0, int c = 0, int d = 0) : first(a), second(b), opposite_first(c), opposite_second(d) {
         isValid = true;
     }
     /// @brief flips the edge by swapping its endpoints with the opposite endpoints
